backend/requirements.txt
# EduVoice Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# AI/ML
google-generativeai>=0.3.0
faster-whisper>=0.10.0

# Text-to-Speech (Free, Unlimited)
edge-tts>=6.1.9

# HTTP Client
httpx>=0.26.0

# Data Validation
pydantic>=2.5.0

# Environment
python-dotenv>=1.0.0

# File Upload Support
python-multipart>=0.0.6
